Add command line flags and simplify some logic
This simplifies the decision on when to change colors for the marble. It introduces a frequency which is used when dropping the marble to decide whether to change the color.

It also removes the transparency effect. It was unused.

Additionally, this commit does a few cleanups:

- Add a new operator<< for ppx::uint3
- Adds command-line arguments for all the fields in SimulationConfig.
- Adds flags for enabling the marble, how frequently it drops and how frequently to do random splats.
- Tidies types for resolution fields.
- Adds command line flags: Remove #ifdef'd out code.  Add command line flags for every configuration setting and also show the compute textures if --enable-debug is used.

+++ projects/fluid_simulation/shaders.cpp
@@ -188,7 +188,7 @@ void ComputeDispatchRecord::BindOutputTexture(uint32_t bindingSlot)
 
 void ComputeShader::Dispatch(const PerFrame& frame, const std::unique_ptr<ComputeDispatchRecord>& dr)
 {
-    ppx::float3 dispatchSize = ppx::float3(dr->mOutput->GetWidth(), dr->mOutput->GetHeight(), 1);
+    ppx::uint3 dispatchSize = ppx::uint3(dr->mOutput->GetWidth(), dr->mOutput->GetHeight(), 1);
 
     PPX_LOG_DEBUG("Running compute shader '" << mShaderFile << ".cs' (" << dispatchSize << ")\n");
